requirements.txt
PyQt5
requests
cryptography
websocket-client